Refactoring, added TODOs
