+++ analytics_dashboard/frontend/src/lib/utils/mappers.ts
